M3: WASM binding + cross-platform byte-identical consistency (verified)
- pqt-wasm: wasm-bindgen API (keypair/encapsulate/decapsulate) exposing the one
  Rust core to JS; randomness passed in by the caller (no in-WASM entropy).
  Builds for wasm32-unknown-unknown; host test reproduces the shared vector.
- Cross-platform consistency oracle (bindings/shared-test-vectors.json) now
  VERIFIED across faces, byte-for-byte:
    * C ABI  — ffi_matches_shared_vector (cargo test -p pqt-ffi)
    * Swift  — swift test passes (Swift -> C ABI -> Rust core)
    * WASM   — decapsulate_matches_shared_vector + wasm32 build
    * Kotlin — wrapper/test written; runs in CI (JDK 22 FFM), not on dev host
- Swift wrapper fixed (uintptr_t -> UInt) and confirmed working with the linked
  static lib.
- CI: bindings-wasm (wasm32 build), bindings-swift (macOS swift test),
  bindings-kotlin (JDK 22 + Gradle) jobs.

32 tests green; clippy -D warnings clean; fmt clean; wasm32 + no_std builds pass.

+++ bindings/swift/Sources/PQTHybrid/PQTHybrid.swift
@@ -21,7 +21,8 @@ public enum PQTHybrid {
     public static func mlkem768Keypair(seed: [UInt8]) throws -> (sk: [UInt8], pk: [UInt8]) {
         var sk = [UInt8](repeating: 0, count: mlkemSkLen)
         var pk = [UInt8](repeating: 0, count: mlkemPkLen)
-        let rc = pqt_mlkem768_keypair(seed, seed.count, &sk, sk.count, &pk, pk.count)
+        let rc = pqt_mlkem768_keypair(
+            seed, UInt(seed.count), &sk, UInt(sk.count), &pk, UInt(pk.count))
         guard rc == PQT_OK else { throw PQTError(code: rc) }
         return (sk, pk)
     }
@@ -29,7 +30,8 @@ public enum PQTHybrid {
     public static func x25519Keypair(secret: [UInt8]) throws -> (sk: [UInt8], pk: [UInt8]) {
         var sk = [UInt8](repeating: 0, count: x25519Len)
         var pk = [UInt8](repeating: 0, count: x25519Len)
-        let rc = pqt_x25519_keypair(secret, secret.count, &sk, sk.count, &pk, pk.count)
+        let rc = pqt_x25519_keypair(
+            secret, UInt(secret.count), &sk, UInt(sk.count), &pk, UInt(pk.count))
         guard rc == PQT_OK else { throw PQTError(code: rc) }
         return (sk, pk)
     }
@@ -42,10 +44,10 @@ public enum PQTHybrid {
     ) throws -> [UInt8] {
         var secret = [UInt8](repeating: 0, count: secretLen)
         let rc = pqt_hybrid_decapsulate(
-            profile.rawValue, suiteId, suiteId.count, policyVersion,
-            skPq, skPq.count, ctPq, ctPq.count, pkPq, pkPq.count,
-            skTrad, skTrad.count, ctTrad, ctTrad.count, pkTrad, pkTrad.count,
-            context, context.count, &secret, secret.count)
+            profile.rawValue, suiteId, UInt(suiteId.count), policyVersion,
+            skPq, UInt(skPq.count), ctPq, UInt(ctPq.count), pkPq, UInt(pkPq.count),
+            skTrad, UInt(skTrad.count), ctTrad, UInt(ctTrad.count), pkTrad, UInt(pkTrad.count),
+            context, UInt(context.count), &secret, UInt(secret.count))
         guard rc == PQT_OK else { throw PQTError(code: rc) }
         return secret
     }
